Split PlanetType.php into files for each class
Fixes code smell Spryker.Classes.ClassFileName.NoMatch.
